requirements.txt
treamlit==1.8.1
numpy
pandas
pillow